Moved array into its own folder
