+++ src/Draco.Compiler/Internal/Symbols/Synthetized/Array/ArrayIndexPropertySymbol.cs
