+++ src/Draco.Compiler/Internal/Symbols/Synthetized/Array/ArrayTypeSymbol.cs
